requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
django-crontab==0.7.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
PyJWT==1.7.1
pytz==2022.1
sqlparse==0.4.2
